Set VM's time based on host's time exposed by Xen
Sets the VM's time when XenIface starts and on resume from suspend.
Uses the underlying host's time, accessed via Xen. Provide an option
for the guest to treat Xen's time as UTC or the current timezone.
This option adds 2 new brandable variables that specify a partial
registry path and value in the VM that determines how the time offset
is modified. These values are REG_TOOLS_PATH and REG_UTC_NAME which
should be specified in the build environment if the values need to
be different.

Signed-off-by: Owen Smith <owen.smith@citrix.com>

Changed REG_TOOLS_PATH to REG_KEY_NAME, which now defaults to 'Windows
PV Drivers' (that being the official XenProject sub-project name), and
use hard coded 'HostTime' key name instead of brandable REG_UTC_NAME.
If we have a brandable parent registry key, then we don't really need
to brand the value names under it.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>