xenagent: get event logging working properly
The xenagent build does not create the necessary DLL for event logging. This
patch re-works the build vcxproj files to create the DLL and adds the
necessary code in package INF file to install it and configure it.

As part of the work the service display name has been changed to match the
module name.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>
Cc: Owen Smith <owen.smith@citrix.com>