Try to make sure the agent always always handles shutdown events
It seems that on Server 2008 the agent does not reliably wake up on
shutdown watch events. This does not seem to occur on any other OS so
it is likely that this is a bug in Server 2008.

To work around the problem this patch nodifies the agent wake up once a
minute and check for a shutdown event (or a suspend event, for
completeness) and act accordingly.

Also this patch squashes the uninteresting error messages that occur
when an attempt is made to read a non-existent xenstore key.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>