Add log message if Windows Update requires a reboot
When Windows requires a reboot to install updates, the reboot and shutdown
operations can take a long time. The agent will have acknowledged the
request, but the operation may not complete in a suitable timeout period.
Add a log entry to help indicate an additional unknown amount of time
will be required for shutdown/reboot.

Signed-off-by: Owen Smith <owen.smith@citrix.com>