Set local time to UTC if Windows has "RealTimeIsUniversal"
Commit 3b8723b1 "Set VM's time based on host's time exposed by Xen"
modified the guest agent to adjust Xen time to UTC before setting the
system time (i.e. the emulated RTC) if a vendor specific registry value
was set.

Windows' idea of whether the RTC is programmed in UTC is actually
controlled by:

[HKLM\System\CurrentControlSet\Control\TimeZoneInformation]
DWORD:RealTimeIsUniversal

If this value is present and non-zero then the RTC is set in UTC otherwise
it is in local time. So, there is no need for a vendor specific key.

This patch removes use of the vendor specific key and checks
"RealTimeIsUniversal" directly to determine how to set the system time.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>